+++ engine/apps/base/utils.py
@@ -4,6 +4,7 @@ from urllib.parse import urlparse
 
 import phonenumbers
 from django.apps import apps
+from django.conf import settings
 from phonenumbers import NumberParseException
 from telegram import Bot
 from twilio.base.exceptions import TwilioException
@@ -136,6 +137,9 @@ class LiveSettingValidator:
 
     @classmethod
     def _check_telegram_webhook_host(cls, telegram_webhook_host):
+        if settings.TELEGRAM_LONG_POLLING_ENABLED:
+            return
+
         try:
             # avoid circular import
             from apps.telegram.client import TelegramClient