+++ CHANGELOG.md
@@ -208,6 +208,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Allow mobile app to consume "internal" schedules API endpoints by @joeyorlando ([#2109](https://github.com/grafana/oncall/pull/2109))
 - Add inbound email address in integration API by @vadimkerr ([#2113](https://github.com/grafana/oncall/pull/2113))
+- Use Telegram polling protocol instead of a webhook if `TELEGRAM_LONG_POLLING_ENABLED` set to `True` by @alexintech
 
 ### Changed